[#51 & #53] - converting mui links to react NavLinks, this stops unnecessary re-renders of the page slowing down performance, adding in a new active state and fixing footer, somewhat [cf]
