+++ src/index.css
@@ -14,6 +14,5 @@
 
 body {
   margin: 0;
-  min-height: 100vh;
 }
